+++ src/hint_processor/builtin_hint_processor/hint_code.rs
@@ -1020,5 +1020,12 @@ remainder_split = split(remainder, num_bits_shift=128, length=2)
 ids.remainder.low = remainder_split[0]
 ids.remainder.high = remainder_split[1]"#;
 
+pub const EC_RECOVER_SUB_A_B: &str = r#"from starkware.cairo.common.cairo_secp.secp_utils import pack
+from starkware.python.math_utils import div_mod, safe_div
+
+a = pack(ids.a, PRIME)
+b = pack(ids.b, PRIME)
+value = res = a - b"#;
+
 #[cfg(feature = "skip_next_instruction_hint")]
 pub const SKIP_NEXT_INSTRUCTION: &str = "skip_next_instruction()";